fixed language changing after submitting form and disabled generating a primary chart in the preview chart window

+++ app/routes.py
@@ -1,3 +1,5 @@
+import os
+
 from flask import render_template, redirect, request 
 from app import app, db
 from app.models import Database
@@ -8,7 +10,8 @@ from app.values import init_values
 @app.route('/', methods=['POST', 'GET'])
 def page_default():
     # to make sure that the language.js file exists
-    open('app/static/javascript/_language.js', 'w+', encoding='utf-8').write(f"var chosen_language = 'EN';")
+    if not os.path.isfile('app/static/javascript/_language.js'):
+        open('app/static/javascript/_language.js', 'w+', encoding='utf-8').write(f"var chosen_language = 'EN';")
     if request.method == 'POST':
         # if submit in the language form is pressed, change the value in the language.js file
         if 'language' in request.form:
@@ -34,7 +37,6 @@ def page_default():
     values = inputs[-1].__dict__
     names = eval('{' + ''.join([line.strip() for line in open('app/static/javascript/translation.js', encoding='utf-8').readlines()][2:-3]) + '}')
     language = open('app/static/javascript/_language.js', 'r', encoding='utf-8').readlines()[0].split()[-1].strip(';').strip("'")
-    print(language)
     
     # creates a file to hold the database content for JS
     open('app/static/javascript/_database.js', 'w+', encoding='utf-8').close()
@@ -42,7 +44,7 @@ def page_default():
     js_db.write("var db = {\n")
     for line in values:
         if line != '_sa_instance_state':
-            js_db.write(f"\t{ line }: { [inputs[i].__dict__[line] for i in range(inputs.count())] },\n")
+            js_db.write(f"\t{ line }: { [inputs[i].__dict__[line] for i in range(1, inputs.count())] },\n")
     js_db.write("}\n")
     js_db.close()
         
@@ -57,7 +59,7 @@ def page_default():
 
 # clears the database on enter
 @app.route('/clear')
-def page_eo():
+def page_clear():
     db.session.query(Database).delete()
     db.session.commit()
     return redirect('/')
